+++ asset_server_go/src/go.mod
@@ -2,13 +2,16 @@ module main
 
 go 1.21.6
 
-replace module/asset => ./asset
-
-require module/asset v0.0.0-00010101000000-000000000000
+require capstone.com/module/tcp v0.0.0
 
 require (
+	capstone.com/module/asset v0.0.0 // indirect
+	capstone.com/module/db v0.0.0 // indirect
 	github.com/go-sql-driver/mysql v1.7.1 // indirect
-	module/db v0.0.0-00010101000000-000000000000 // indirect
 )
 
-replace module/db => ./db
+replace capstone.com/module/asset => ./asset
+
+replace capstone.com/module/db => ./db
+
+replace capstone.com/module/tcp => ./tcp
